+++ test-suite/conftest.py
@@ -67,18 +67,19 @@ async def config():
     yield config
 
     # Cleanup
-    await wallet.close_wallet(config.wallet_handle)
-    await wallet.delete_wallet(
-        json.dumps({
-            'id': config.wallet_name,
-            'storage_config': {
-                'path': config.wallet_path
-            }
-        }),
-        json.dumps({'key': 'test-agent'})
-    )
+    if config.clear_wallets:
+        await wallet.close_wallet(config.wallet_handle)
+        await wallet.delete_wallet(
+            json.dumps({
+                'id': config.wallet_name,
+                'storage_config': {
+                    'path': config.wallet_path
+                }
+            }),
+            json.dumps({'key': 'test-agent'})
+        )
 
-    os.rmdir(config.wallet_path)
+        os.rmdir(config.wallet_path)
 
 
 @pytest.fixture(scope='session')
